requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
httpx==0.25.1
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.4.2
pandas==2.1.1
numpy==1.26.1 